renamed a class and deleted a duplicate
