+++ Anti-Fraud System/task/src/antifraud/model/response/TransactionResponse.java
@@ -2,5 +2,9 @@ package antifraud.model.response;
 
 import java.util.List;
 
-public record TransactionResponse(String result, List<String> info) {
+public record TransactionResponse(String result, String info) {
+
+    public TransactionResponse(String result, List<String> info) {
+        this(result, String.join(", ", info));
+    }
 }
